refactor: restructure package
Separate asymmetric and symmetric cryptography into subpackages, and rename files to reflect the
underlying algorithms they use. This will make it easier to add onto the library. Also, change package name.

BREAKING CHANGE: Names of objects and files have changed and they are now in subclasses, and the package name has changed. Functionality is not changed.