feat: add support for multiple recipients
write methods to encrypt a message for multiple recipients without massively increasing ciphertext
size, modify existing methods to work with the slight modification to the message format (first byte
now represents the number of recipient keys so the decryptor can know where to look), add an
override for encryptAndSign to handle a list of keys

BREAKING CHANGE: new message format, which consists of an extra byte at the start of the message to
indicate the number of recipients. Messages from previous versions of REC can be decrypted with
legacyDecrypt(message, key). This method will be removed with the next major version.

complete #1