fix: fix incorrect math
encrypt method for multiple keys previously did not obey the intended behavior of a zero byte
indicating 1 recipient, a 1 byte indicating 2 recipients, etc. The reason for this design choice is
that zero recipients is impossible so really the first byte represents the number of extra
recipients after the first mandatory one, not the total number of recipients.