feat: add optional config to encryption
add default-valued-parameters (overrides in Java) to allow users to configure REC if they would
like. These will not be documented very much because they're meant for people who know what they're
doing. The parameters include things like the AES keysize used by the RSAHandler's encryption
methods, and the algorithm used for ciphers and signatures. Most users can simply ignore these
options since the defaults follow best practice, but if for any reason someone needs for example to
use a different type of padding they can. I will not be updating the code to support non-GCM AES at
this time since that would be tedious but it may happen in the future.